Update made to stabalize server play with current Release candidate.
I am trying to commit this change in order for us to be able to connect to a server. By doing something this simple, it will essentially keep all code and be easy to fix later, and it will (Hopefully -_-) solve the issue of the cape render errors upon connection.